[CHORE] #12 - apply dp to px function
