+++ app/src/main/java/org/mascota/ui/view/home/data/datasource/LocalHomeBookDataSource.kt
@@ -7,5 +7,5 @@ import org.mascota.ui.view.home.data.model.HomePageInfoData
 class LocalHomeBookDataSource : HomeBookDataSource {
     override fun getHomeBookInfoData(): HomeBookInfoData = HomeBookInfoData("1부 ", "모여봐요 코봉의 숲", "https://cdn.pixabay.com/photo/2014/04/13/20/49/cat-323262_960_720.jpg")
     override fun getHomeDiaryInfoData(): HomeDiaryInfoData = HomeDiaryInfoData("162화", "제목", "내용", "2020.02.20")
-    override fun getHomePageInfoData(): HomePageInfoData = HomePageInfoData(prolog = true)
+    override fun getHomePageInfoData(): HomePageInfoData = HomePageInfoData(prolog = false)
 }
