+++ app/src/main/java/org/mascota/util/BindingAdapter.kt
@@ -1,9 +1,11 @@
 package org.mascota.util
 
+import android.annotation.SuppressLint
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
 import org.mascota.R
 import org.mascota.ui.view.calendar.CalendarFragment.Companion.DOG_ANGRY
 import org.mascota.ui.view.calendar.CalendarFragment.Companion.DOG_BORING
@@ -62,5 +64,18 @@ object BindingAdapter {
                 else -> View.GONE
             }
         }
+    }
+
+    @JvmStatic
+    @BindingAdapter("imgUrl")
+    fun setRemoteCircleImage(image: ImageView, url: String?) {
+        Glide.with(image.context).load(url).into(image)
+    }
+
+    @SuppressLint("SetTextI18n")
+    @JvmStatic
+    @BindingAdapter("TitleText")
+    fun setEpisodeText(textView: TextView, number: Int?) {
+        textView.text = number.toString() + "화 기록하기"
     }
 }
